Rearrange test util methods to make them available in Sockeye library
